requirements.txt
requests
flask
mysql-connector-python
flask-cors
qrcode
bcrypt
flask_jwt_extended